refactor: moved getFileHash function into its own helper function
